requirements.txt
stop-words
beautifulsoup4
googlesearch-python